FIXED SINGLETONS
changed singletons to get singleton static instance, and .Instance, to get teh one we want, instead of searching for it.  Turns out Destroy waits until end of update, so all of Starts() had ahppened and found the bad singleton.  Also added a set level function, defaulted it to lvl 2, and have lvl 1 set it to new.