Flame mortar tower now works, still touch up.
Mortar tower kicks ass.  It shoots, explodes, leaves an AOE.  Seems cool and fun.  It might shoot a little outside of its attack range?  and the head needs to me changed when looking at an enemy, so it only spins on a single axis.  Need to also make it scales the AOE size.